# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FocusQuest API - ADHD gamified task tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/me/streak-freeze/buy:
    post:
      operationId: buyStreakFreeze
      tags: [users]
      summary: Spend 50 XP to buy a streak freeze (max 1 held at a time)
      responses:
        "200":
          description: Freeze purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakFreezeResult"
        "400":
          description: Not enough XP or already holding a freeze
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /users/me/xp-history:
    get:
      operationId: getMyXpHistory
      tags: [users]
      summary: Get daily XP earned for the past N days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: XP history per day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/XpDataPoint"
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search for users by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
components:
...
  schemas:
...
    User:
      type: object
      required: [id, username, totalPoints, currentLevel, streakDays, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarColor:
          type: string
        totalPoints:
          type: integer
        weeklyPoints:
          type: integer
        currentLevel:
          type: integer
        levelName:
          type: string
        streakDays:
          type: integer
        longestStreak:
          type: integer
        pointsToNextLevel:
          type: integer
        createdAt:
          type: string
    UserSummary:
      type: object
      required: [id, username, currentLevel, totalPoints]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarColor:
          type: string
        currentLevel:
          type: integer
        levelName:
          type: string
        totalPoints:
          type: integer
        streakDays:
          type: integer
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        avatarColor:
          type: string
...
    StreakFreezeResult:
      type: object
      required: [streakFreezes, totalPoints]
      properties:
        streakFreezes:
          type: integer
        totalPoints:
          type: integer
...
    XpDataPoint:
      type: object
      required: [date, label, xp]
      properties:
        date:
          type: string
          description: "YYYY-MM-DD"
        label:
          type: string
          description: "Short weekday label e.g. Mon"
        xp:
          type: integer
...